Separate DynamicSystem and Brain Objects
