+++ brainpy/integrators/sde/__init__.py
@@ -6,6 +6,6 @@ Numerical methods for stochastic differential equations.
 
 from .euler_and_milstein import *
 from .srk_scalar import *
-from .srk_strong import *
+# from .srk_strong import *
 # from .srk_weak import *
