+++ brainpy/simulation/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 
-from .network import *
-from .population import *
+from .brain_objects import *
+
