+++ brainpy/simulation/utils.py
@@ -173,14 +173,14 @@ def format_net_level_inputs(inputs, run_length):
     formatted_input : dict
         The formatted input.
     """
-    from brainpy.simulation import population
+    from brainpy.simulation import brain_objects
 
     # 1. format the inputs to standard
     #    formats and check the inputs
     if not isinstance(inputs, (tuple, list)):
         raise errors.ModelUseError('"inputs" must be a tuple/list.')
     if len(inputs) > 0 and not isinstance(inputs[0], (list, tuple)):
-        if isinstance(inputs[0], population.Population):
+        if isinstance(inputs[0], brain_objects.DynamicSystem):
             inputs = [inputs]
         else:
             raise errors.ModelUseError('Unknown input structure. Only supports '
@@ -199,7 +199,7 @@ def format_net_level_inputs(inputs, run_length):
     formatted_inputs = {}
     for input in inputs:
         # target
-        if isinstance(input[0], population.Population):
+        if isinstance(input[0], brain_objects.DynamicSystem):
             target = input[0]
             target_name = input[0].name
         else:
